src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlparse
pandas
networkx
langchain
langchain-community
ollama
python-dotenv
rich
python-multipart
pyyaml 